AOT-718 refactoring duplicates, TODOs and removing redundant code for reading Nexus secret
